+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [unreleased]
 
+# [v0.2.0] 2024-06-28
+
+## Fixed
+
+- The large file flag was not set properly in the source handler for large file transfers.
+
+## Changed
+
+- Added `file_size` abstract method to `VirtualFilestore`
+- Renamed `HostFilestore` to `NativeFilestore`, but keep old name alias for backwards compatibility.
+
 # [v0.1.2] 2024-06-04
 
 Updated documentation configuration to include a `spacepackets` docs mapping. This